fix: predict_linear_poly の標準化・積演算をfloat32二重丸めから倍精度に統一
C++版(type4=linear_poly)の (x-center)/scale 標準化と積・二乗項 s[a]*s[b] を
float単精度のまま計算すると、JS版(predict-core.js、全体を倍精度で計算してから
1回だけfloat32へ丸める)と極端な入力値(x=1e6等)で丸め順序の差が無視できない絶対誤差
(round_output後で~1e3〜1e4オーダー)として残ることが、type4パリティフィクスチャ追加の
過程で判明した。predict_linear/predict_mlpと同様に「倍精度で計算→最後に1回だけfloatへ」
に統一し、JS側と丸めのタイミングを揃える。既存40フィクスチャのmatrix_cpp_outには
ドリフトなし(影響範囲はtype4のみ)。

.gitignore: C++参照実装の検証用ビルド(predict_native_ref/.exe)がpredict-parity.ymlで
毎回生成される割に追跡除外されていなかったため追加。

+++ native_predictor/predict_native_v2.cpp
@@ -502,14 +502,29 @@ static float predict_mlp(const MLPModel& m, const float* x, int d) {
 // (train_bridge._write_treg_stream / predict-core.js predictLinearPoly と同一ロジック)。
 static float predict_linear_poly(const LinearPolyModel& m, const float* x, int d) {
     std::vector<float> s(d);
+    // 高-N2(2026-07-16): float同士の減算→除算を素朴に行うと、2回の丸め(減算の丸め・
+    // 除算の丸め)が積み重なる。JS版(predict-core.js predictLinearPoly)は
+    // x[i]/center[i]/scale[i]をいずれもJS数値(=倍精度で保持された値)として扱い、
+    // (x[i]-center[i])/scale[i] 全体を倍精度で計算してから Float32Array 代入時に
+    // 1回だけ丸める。極端な入力値(例: x=1e6 と小さいcenter/scaleの組)ではこの
+    // 丸め回数の差が無視できない誤差になる(linear_poly_derived_*フィクスチャで
+    // 実測、round_output=Trueの丸め後比較で~3.8e3の不一致として顕在化)。
+    // predict_linear/predict_mlp と同様に「倍精度で計算→最後に1回だけfloatへ」に
+    // 統一し、JS側と丸めのタイミングを揃える。
     for (int i = 0; i < d; i++)
-        s[i] = (x[i] - m.center[i]) / m.scale[i];
+        s[i] = (float)(((double)x[i] - (double)m.center[i]) / (double)m.scale[i]);
 
     double sum = m.intercept;
     const size_t n = m.coef.size();
     for (size_t t = 0; t < n; t++) {
         int32_t a = m.term_a[t], b = m.term_b[t];
-        float val = (b < 0) ? s[a] : s[a] * s[b];
+        // 高-N2(2026-07-16): 積・二乗項 s[a]*s[b] を float(単精度) で計算してから
+        // coef との積に回すと、単精度どうしの積が丸められてから加算されるため、
+        // JS側(predict-core.js predictLinearPoly、s[a]/s[b]はFloat32Arrayだが積は
+        // JS数値=倍精度で計算される)と極端な入力値(標準化後の値が大きくなる列)で
+        // 有意な不一致が生じることが type4 パリティフィクスチャ(linear_poly_derived_*)の
+        // 追加で判明した。積演算自体を double に上げてJS側と同一の丸め経路にする。
+        double val = (b < 0) ? (double)s[a] : (double)s[a] * (double)s[b];
         sum += (double)m.coef[t] * val;
     }
     return (float)sum;
